Primitive enhancements and optimisations

+++ convex-core/src/main/java/convex/core/data/prim/CVMLong.java
@@ -288,7 +288,7 @@ public final class CVMLong extends AInteger {
 	}
 
 	@Override
-	public long byteLength() {
+	public int byteLength() {
 		return Utils.byteLength(value);
 	}
